==> Lethality_and_Deformities/3.1-CurveP_bmd_affected_rate_positive_direction.Rmd <==
---
title: "3.1_CurveP_affected_rate_positive_direction"
author: "Jory Curry"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  html_document:
    code_folding: show
    df_print: paged
    toc: true
    toc_depth: 5
    toc_float: true
    theme: readable
params:
  direction: "pos"
  sample_size: 1000
  n_dose_groups: 6
  endpoint: "Affected.Rate"
---

```{r setup, include=FALSE}
#knitr::opts_chunk$set(echo = TRUE)
```

```{r Libraries, include=FALSE}
library(Rcurvep)
library(tidyverse)
library(future)
library(future.apply)
library(furrr)
library(kableExtra)
source("Functions/RCurveP_Data_Wrangling_Functions.R")
```

```{r import_data, include=FALSE, warning=FALSE, message=FALSE}
ar_dat <- read_csv(file = "Data/apicaldata_ol_rm_and_Lethal_Dose-lowest_dose_removed.csv")

metaData <- read.csv(file = "Data/MetaData.csv", skip = 1) #Import the Meta Data that includes information about the data in the folders
#CAS is the Chemical Abstract Service, MOA is the Mode of Action. This table includes useful information about the exposure concentrations for each chemical dose in mg/L. We'll use this later to create our final data frame
doseData <- metaData %>%
  dplyr::select(Chemical, Dose_1:Dose_6) %>%
  rename(Control = Dose_6) %>%
  gather(key = Dose, value = "Dose_mg_L", Dose_1:Control)

lowest_dose_treatment_group <- doseData %>%
  filter(Dose != "Control") %>%
  dplyr::summarize(lowest_dose_treatment_group = min(Dose_mg_L))
lowest_dose_treatment_group <- lowest_dose_treatment_group[1,][[1]]

ss <- params$sample_size
dir <- params$direction
number_of_dose_groups_per_chem <- params$n_dose_groups # Set this to the number of dose groups you have (including control)

bmr_thresh <- readRDS("Data/positive_direction_bmr.R")
input_tib <- readRDS("Data/positive_direction_bmr_tibble.R")
```

```{r prep_data_for_rCurveP, echo=TRUE}
ar_dat <- ar_dat %>%
  mutate(endpoint = "Affected.Rate") %>%
  mutate(resp = get(params$endpoint)) %>%
  rename(chemical = Chemical) %>%
  mutate(Dose..mg.L. = Dose..mg.L./lowest_dose_treatment_group[[1]]) %>%
  mutate(conc = log10(Dose..mg.L.)) %>%
  group_by(chemical, endpoint) %>%
  mutate(control_conc = max(conc)-(number_of_dose_groups_per_chem-1)) %>% #making sure there are no infinite values for concentrations. Controls are made to be 1 dose group down from the lowest concentration
  mutate(conc = if_else(condition = conc == -Inf, true = control_conc, false = conc)) %>%
  dplyr::select(endpoint, chemical, conc, resp)
```

```{r CurveP_calc, echo=TRUE}
#calculate the BMD step 1
system.time(
  bmd_dat_pos <-
    future_pmap(
      input_tib,
      ~ combi_run_rcurvep(
        ..4,
        TRSH = ..3,
        RNGE = ..1,
        MXDV = (1*0.05), #Default value is 5/100 so adjusting scale... Maximum allowed deviation from monotonicity (default = 5).
        n_samples = ss,
        keep_sets = c("act_set", "resp_set"),
        TrustHi = FALSE,
        seed = 42069
      ),
      .options = furrr_options(seed = 42069)
    )
)
```

```{r summarise_rcurveP_results, echo=TRUE, message=FALSE, warning=FALSE}
#calculate the BMD step 2
sum_bmd_dat_pos <-
  summarize_rcurvep_output(
    bmd_dat_pos[[1]],
    inactivate = "INVERSE",
    ci_level = 0.95
  )

#summarise the results
simulated_curves_summary <-
  summarise_rcurvep_results(
    resp_set = sum_bmd_dat_pos$result$resp_set,
    act_set = sum_bmd_dat_pos$result$act_set,
    act_summary = sum_bmd_dat_pos$act_summary,
    reject_hit_conf_under = 0.001
  )

bmds <- bmd_results(simulated_curves_summary)

curve_hits <- simulated_curves_summary %>%
  filter(hit == 1)

confident_curves <- confident_hits(summary_dat = simulated_curves_summary, reject_hit_conf_under = 0.001)
# If Error says "out of bounds", reduce conf filter
```

```{r _simulated_plot_results, echo=FALSE, warning=FALSE}
#Data Wrangling for nice labels
simulated_curves_summary <- simulated_curves_summary %>%
  group_by(chemical) %>%
  mutate(x_label_num = if_else(condition = conc == lowest_conc, true = 0, false = (10^conc)*lowest_dose_treatment_group)) %>%
  mutate(x_label = factor(x_label_num, levels = c(0, lowest_dose_treatment_group, lowest_dose_treatment_group*10, lowest_dose_treatment_group*100, lowest_dose_treatment_group*1000, lowest_dose_treatment_group*10000, lowest_dose_treatment_group*100000, lowest_dose_treatment_group*1000000, lowest_dose_treatment_group*10000000, lowest_dose_treatment_group*100000000, lowest_dose_treatment_group*1000000000, lowest_dose_treatment_group*10000000000), ordered = TRUE))

curve_hits <- curve_hits %>%
  group_by(chemical) %>%
  mutate(x_label_num = if_else(condition = conc == lowest_conc, true = 0, false = (10^conc)*lowest_dose_treatment_group)) %>%
  mutate(x_label = factor(x_label_num, levels = c(0, lowest_dose_treatment_group, lowest_dose_treatment_group*10, lowest_dose_treatment_group*100, lowest_dose_treatment_group*1000, lowest_dose_treatment_group*10000, lowest_dose_treatment_group*100000, lowest_dose_treatment_group*1000000, lowest_dose_treatment_group*10000000, lowest_dose_treatment_group*100000000, lowest_dose_treatment_group*1000000000, lowest_dose_treatment_group*10000000000), ordered = TRUE))



#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#

p <- ggplot(NULL, aes(x_label, resp)) +
  geom_point(
    data = simulated_curves_summary,
    mapping = aes(x = x_label, y  = resp, group = sample_id),
    color = "gray75",
    alpha = 0.2
  ) +
  geom_point(
    data = curve_hits,
    mapping = aes(x = x_label, y = resp, group = sample_id),
    color = "#a9a9a9",
    alpha = 0.7
  ) +
  geom_line(
    data = simulated_curves_summary,
    mapping = aes(x = x_label, y = resp, group = sample_id),
    color = "gray75",
    alpha = 0.2
  ) +
  geom_line(
    data = curve_hits,
    mapping = aes(x = x_label, y = resp, group = sample_id),
    color = "#a9a9a9",
    alpha = 0.7
  ) +
  facet_wrap( ~ chemical, scales = "free_x") +
  ylim(-0.3, 1) +
  theme_minimal() +
  theme(
    panel.border = element_rect(colour = "grey50", fill = NA),
    #panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8,1)) +
  xlab("Dose (mg/L)") +
  ylab("Affected Rate")
```

```{r save_the_resutls, include=FALSE}
#save the image
ggplot2::ggsave(
  plot = p,
  filename = paste0(
    "Output/Images/", "affected_rate_", dir, "_", "_simulated_curves_",
    ss,
    "_samples.png"
  ),
  device = "png",
  units = "px",
  width = 1920,
  height = 1080,
  bg = "white",
  scale = 3
)
```

```{r print_image, echo=FALSE}
knitr::include_graphics(
  paste0(
    "Output/Images/",
    "affected_rate_",
    dir,
    "_",
    "_simulated_curves_",
    ss,
    "_samples.png"
  )
)
```

```{r processed_results, echo=FALSE}
  
# p_facet <- ggplot() +
#   geom_point(
#     data = simulated_curves_summary,
#     mapping = aes(x = conc, y  = corrected_resp, group = sample_id),
#     color = "gray75",
#     alpha = 0.2
#   ) +
#   geom_point(
#     data = curve_hits,
#     mapping = aes(x = conc, y = corrected_resp, group = sample_id),
#     color = "#a9a9a9",
#     alpha = 0.7
#   ) +
#   geom_line(
#     data = simulated_curves_summary,
#     mapping = aes(x = conc, y = corrected_resp, group = sample_id),
#     color = "gray75",
#     alpha = 0.2
#   ) +
#   geom_line(
#     data = curve_hits,
#     mapping = aes(x = conc, y = corrected_resp, group = sample_id),
#     color = "#a9a9a9",
#     alpha = 0.7
#   ) +
#   # geom_line(data = confident_curves[["cil"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "purple") +
#   # geom_line(data = confident_curves[["ciu"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "yellow") +
#   # geom_line(data = confident_curves[["mean"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "green") +
#   # geom_line(data = confident_curves[["median"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "blue") +
#   geom_hline(
#     yintercept = bmr_thresh,
#     linetype = "dashed",
#     color = "red"
#   ) +
#   geom_vline(data = bmds,
#              aes(xintercept = median_POD),
#              color = "blue") +
#   # geom_vline(data = bmds,
#   #            aes(xintercept = POD_med),
#   #            color = "green") +
#   geom_vline(data = bmds, aes(xintercept = ciu_POD), color = "yellow") +
#   # geom_vline(data = bmds_pearson, aes(xintercept = mean_POD), color = "green") +
#   geom_vline(data = bmds, aes(xintercept = cil_POD), color = "purple") +
#   geom_text(
#     data = bmds,
#     mapping = aes(label = paste("Hit =", hit_confidence), x = lowest_conc),
#     position = position_nudge(x = 1),
#     y = 0.85,
#     inherit.aes = FALSE,
#     size = 4
#   ) +
#   facet_wrap( ~ chemical, scales = "free_x") +
#   ylim(-0.3, 1) +
#   theme_minimal() +
#   theme(
#     panel.border = element_rect(colour = "grey50", fill = NA),
#     #panel.grid.major = element_blank(),
#     panel.grid.minor = element_blank()
#   ) +
#   scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8,1)) +
#   scale_x_continuous(breaks = c(-1, 0, 1, 2, 3, 4, 5, 6 , 7, 8, 9, 10)) +
#   xlab("Dose log10(mg/L)*1e-7") +
#   ylab("Corrected Affected Rate")
```

```{r processed_results_No_facet, echo=FALSE, warning=FALSE}
options(scipen2 = 3)
print.numeric <- function(x, ...) {
  scipen2 <- getOption("scipen2", default = 3)
  ifelse(abs(log(x, 10L)) >= scipen2,
               format(x, digits = 5, scientific = TRUE),
               x)
}
chemicalNames <- unique(simulated_curves_summary$chemical)
p_list <- list()
for (k in 1:length(chemicalNames)) {
  labels <- simulated_curves_summary %>%
    filter(chemical == chemicalNames[k]) %>%
    select(x_label_num) %>%
    pull() %>%
    unique() %>%
    print.numeric()
  temp_simulated <- simulated_curves_summary %>%
    filter(chemical == chemicalNames[k])
  temp_curve_hits <- curve_hits %>%
    filter(chemical == chemicalNames[k])
  temp_bmds <- bmds %>%
    filter(chemical == chemicalNames[k])
  
  p <- ggplot() + 
    geom_point(
    data = temp_simulated,
    mapping = aes(x = conc, y  = corrected_resp, group = sample_id),
    color = "gray75",
    alpha = 0.2
  ) +
  geom_point(
    data = temp_curve_hits,
    mapping = aes(x = conc, y = corrected_resp, group = sample_id),
    color = "#a9a9a9",
    alpha = 0.7
  ) +
  geom_line(
    data = temp_simulated,
    mapping = aes(x = conc, y = corrected_resp, group = sample_id),
    color = "gray75",
    alpha = 0.2
  ) +
  geom_line(
    data = temp_curve_hits,
    mapping = aes(x = conc, y = corrected_resp, group = sample_id),
    color = "#a9a9a9",
    alpha = 0.7
  ) +
  # geom_line(data = confident_curves[["cil"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "purple") +
  # geom_line(data = confident_curves[["ciu"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "yellow") +
  # geom_line(data = confident_curves[["mean"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "green") +
  # geom_line(data = confident_curves[["median"]], mapping = aes(x = conc, y = corrected_resp, group = sample_id), color = "blue") +
  geom_hline(
    yintercept = bmr_thresh,
    linetype = "dashed",
    color = "red"
  ) +
  {if(temp_bmds$ciu_POD != temp_bmds$highest_conc)geom_vline(data = temp_bmds, aes(xintercept = ciu_POD), color = "black")} +
  # geom_vline(data = bmds_pearson, aes(xintercept = mean_POD), color = "green") +
  {if(temp_bmds$cil_POD != temp_bmds$highest_conc)geom_vline(data = temp_bmds, aes(xintercept = cil_POD), color = "black")} +
     {if(temp_bmds$median_POD != temp_bmds$highest_conc)geom_vline(data = temp_bmds,
             aes(xintercept = median_POD),
             color = "blue")} +
  geom_text(
    data = temp_bmds,
    mapping = aes(label = paste("Hit =", hit_confidence), x = lowest_conc),
    position = position_nudge(x = 1),
    y = 0.85,
    inherit.aes = FALSE,
    size = 4
  ) +
  theme_minimal() +
  theme(
    panel.border = element_rect(colour = "grey50", fill = NA),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.title.x=element_blank(),
    axis.title.y=element_blank()
  ) +
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8,1), limits = c(0,1)) +
  scale_x_continuous(breaks = unique(temp_simulated$conc), labels = labels) +
  # xlab("Dose (mg/L)") +
  # ylab("Corrected Affected Rate") +
  ggtitle(paste0(chemicalNames[k]))
  
  p_list[[chemicalNames[k]]] <- list(p, chemicalNames[k])
}
```

```{r alternative_save_the_resutls, include=FALSE}
yleft <-
  grid::textGrob("Corrected Affected Rate Response",
           rot = 90,
           gp = grid::gpar(fontsize = 12))
bottom <- grid::textGrob("Dose (mg/L)", gp = grid::gpar(fontsize = 12))
p <- gridExtra::grid.arrange(
  p_list[[1]][[1]],
  p_list[[2]][[1]],
  p_list[[3]][[1]],
  p_list[[4]][[1]],
  p_list[[5]][[1]],
  p_list[[6]][[1]],
  p_list[[7]][[1]],
  p_list[[8]][[1]],
  p_list[[9]][[1]],
  p_list[[10]][[1]],
  p_list[[11]][[1]],
  p_list[[12]][[1]],
  p_list[[13]][[1]],
  p_list[[14]][[1]],
  p_list[[15]][[1]],
  p_list[[16]][[1]],
  p_list[[17]][[1]],
  p_list[[18]][[1]],
  p_list[[19]][[1]],
  p_list[[20]][[1]],
  p_list[[21]][[1]],
  p_list[[22]][[1]],
  p_list[[23]][[1]],
  p_list[[24]][[1]],
  p_list[[25]][[1]],
  p_list[[26]][[1]],
  p_list[[27]][[1]],
  p_list[[28]][[1]],
  p_list[[29]][[1]],
  ncol = 6,
  nrow = 5,
  left = yleft,
  bottom = bottom
)

#save the image
ggsave(
  plot = p,
  filename = paste0(
    "Output/Images/", "affected_rate_", dir, "_", "_processed_curves_",
    ss,
    "_samples.png"
  ),
  device = "png",
  units = "px",
  width = 1920,
  height = 1080,
  bg = "white",
  scale = 3
)
```

```{r alternative_print_image, echo=FALSE}
knitr::include_graphics(
  paste0(
    "Output/Images/",
    "affected_rate_",
    dir,
    "_",
    "_processed_curves_",
    ss,
    "_samples.png"
  )
)
```

```{r alternative_vis, include=FALSE}
#Distribution of BMDs
# p_distribution_bmds <- curve_hits %>%
#   group_by(chemical) %>%
#   ggplot(data = ., aes(x = POD)) +
#   geom_histogram() +
#   facet_wrap( ~ chemical, scales = "free_x") +
#   geom_vline(aes(xintercept = median_POD), color = "blue") +
#   geom_vline(aes(xintercept = cil_POD), color = "purple") +
#   geom_vline(aes(xintercept = ciu_POD), color = "yellow") +
#   geom_vline(aes(xintercept = lowest_conc),
#              color = "black",
#              linetype = "dashed") +
#   geom_vline(aes(xintercept = highest_conc),
#              color = "black",
#              linetype = "solid") +
#   theme_classic() +
#   scale_x_continuous(expand = expansion(add = 1)) +
#   xlab("BMD") +
#   geom_text(
#     mapping = aes(label = paste("Hit =", hit_confidence), x = lowest_conc +
#                     2),
#     y = 1000,
#     inherit.aes = FALSE,
#     size = 4
#   )
# 
# #save the image
# ggsave(
#   plot = p_distribution_bmds,
#   filename = paste0(
#     "Output/Images/", "z_score_", dir, "_", "_distribution_bmds_",
#     ss,
#     "_samples.svg"
#   ),
#   device = "svg",
#   units = "px",
#   width = 1920,
#   height = 1080,
#   bg = "white",
#   scale = 2
# )
```

```{r table_the_results, echo=TRUE}
#Table the results
temp <- list(bmds)
names(temp) <- c(paste0("affected_rate_", dir, "_dir_", ss, "_ss"))
all_summarized_bmd_dat <- plyr::ldply(temp)
all_summarized_bmd_dat <- all_summarized_bmd_dat %>%
  arrange(chemical)

all_summarized_bmd_dat %>%
  mutate(lowest_conc = lowest_conc + 1) %>%
  mutate(
    lowest_conc = (10 ^ lowest_conc)*lowest_dose_treatment_group,
    highest_conc = sprintf((10 ^ highest_conc)*lowest_dose_treatment_group, fmt = '%.7f'),
    cil_POD = (10 ^ cil_POD)*lowest_dose_treatment_group,
    POD_med = (10 ^ POD_med)*lowest_dose_treatment_group,
    ciu_POD = (10 ^ ciu_POD)*lowest_dose_treatment_group,
    median_POD = (10 ^ median_POD)*lowest_dose_treatment_group,
    mean_POD = (10 ^ mean_POD)*lowest_dose_treatment_group
  ) %>%
  mutate(POD_med = if_else(is.na(POD_med), true = median_POD, false = POD_med)) %>%
  dplyr::select(
    chemical,
    .id,
    lowest_conc,
    highest_conc,
    cil_POD,
    POD_med,
    ciu_POD,
    hit_confidence
  ) %>%
  kable(
    col.names = c(
      "Chemical",
      "endpoint",
      "Low Dose (mg/L)",
      "High Dose (mg/L)",
      "BMD - low confidence interval (0.05)",
      "BMD",
      "BMD - high confidence interval (0.95)",
      "Hit Confidence"
    ),
    caption = "Table 9. Summary of all the benchmark doses (BMDs) of every chemical"
  ) %>%
  kable_styling() %>%
  row_spec(
    which(
      all_summarized_bmd_dat$median_POD < all_summarized_bmd_dat$highest_conc
    ),
    bold = TRUE,
    background = "gray"
  ) %>%
  scroll_box()
```

```{r 2.save_the_resutls, include=FALSE}
#Save the data
data <- all_summarized_bmd_dat %>%
  mutate(lowest_conc = lowest_conc + 1) %>%
  mutate(
    lowest_conc = (10 ^ lowest_conc)*lowest_dose_treatment_group,
    highest_conc = sprintf((10 ^ highest_conc)*lowest_dose_treatment_group, fmt = '%.7f'),
    cil_POD = (10 ^ cil_POD)*lowest_dose_treatment_group,
    POD_med = (10 ^ POD_med)*lowest_dose_treatment_group,
    ciu_POD = (10 ^ ciu_POD)*lowest_dose_treatment_group,
    median_POD = (10 ^ median_POD)*lowest_dose_treatment_group,
    mean_POD = (10 ^ mean_POD)*lowest_dose_treatment_group
  ) %>%
  mutate(POD_med = if_else(is.na(POD_med), true = median_POD, false = POD_med))

write_csv(x = data, file = paste0("Output/", "BMDs/", "affected_rate_", dir, "_", "BMDs_", ss, "_samples.csv"))
```
