==> Lethality_and_Deformities/4.2-Save_Images_of_Fitted_bmd_affected_rate_positive_direction.Rmd <==
---
title: "4.1_Fitted_affected_rate_positive_direction"
author: "Jory Curry"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  html_document:
    code_folding: show
    df_print: paged
    toc: true
    toc_depth: 5
    toc_float: true
    theme: readable
params:
  direction: "pos"
  sample_size: 100
  n_dose_groups: 6
  endpoint: "Affected.Rate"
---

```{r setup, include=FALSE}
#knitr::opts_chunk$set(echo = TRUE)
```

```{r Libraries, include=FALSE}
library(Rcurvep)
library(tidyverse)
library(future)
library(future.apply)
library(furrr)
library(kableExtra)
library(tcpl)
library(grid)
library(gridExtra)
library(png)
library(knitr)
source("Functions/RCurveP_Data_Wrangling_Functions.R")
```

```{r load_the_results, include=FALSE}
ss <- params$sample_size
bmr_thresh <- readRDS(file = paste0("Data/positive_direction_bmr.R"))
fitted_output_summary <- readRDS(file = paste0("Data/fitted_output_summary_", ss, "_samples.RDS"))
bmds_fitted <- readRDS(file = paste0("Data/fitted_bmds_", ss, "_samples.RDS"))
```

```{r Fitted_plot_results, echo=FALSE}
chemicalNames <- unique(fitted_output_summary$chemical)
options(scipen2 = 3)
print.numeric <- function(x, ...) {
  scipen2 <- getOption("scipen2", default = 3)
  ifelse(abs(log(x, 10L)) >= scipen2,
               format(x, digits = 5, scientific = TRUE),
               x)
}
#Plotting
#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#
#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#
p_list <- list()
for (k in 1:length(chemicalNames)) {
  temp_simulated <- fitted_output_summary %>%
    filter(chemical == chemicalNames[k])
  temp_bmds <- bmds_fitted %>%
    filter(chemical == chemicalNames[k])

#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#
#ToxCast Pipeline (tcpl) package

#Checking RCurveP HIll Model Fitted responses and hill paramaters... Yep looks exactly the same
p1 <- temp_simulated %>%
  group_by(sample_id) %>%
  #Calculate the baseline response median absolute deviation
  mutate(bmad = mad(resp, center = median(resp))) %>%
  nest() %>%
  ungroup() %>%
  #Fit Hill model and extract estimated paramaters
  mutate(model = map(data, ~ tcplFit(.x$conc, .x$resp, bmad = unique(.x$bmad), bidirectional = TRUE, force.fit = TRUE)))
#Definte fitted response values for plot
p1_point <- plyr::ldply(p1$data)

#Define white space around plot
par(mar = c(2.1,2.1,1,0.1))
#Plot Fitted response values
plot(p1_point$fitted_resp ~ p1_point$conc, pch = 16, axes=FALSE, frame.plot=TRUE, ylim=c(0,1), ann=FALSE)
#add title
title(main = chemicalNames[k])
#Build the X axis
Axis(side = 1, labels=print.numeric(unique(p1_point$x_label_num)), at = unique(p1_point$conc))
Axis(side = 2, labels = c(0,0.2,0.4,0.6,0.8,1), at = c(0,0.2,0.4,0.6,0.8,1))
#Build the Y axis
#plot each hill model using tcpl funstion tcplAddModel
for(i in 1:ss){
#Define what model to plot
modl <- if_else(condition = !is.null(p1[[3]][[i]]$hill_modl), true = "hill", false = if_else(condition = !is.null(p1[[3]][[i]]$gnls_modl), true = "gnls", false = "cnst"))
#Only show hill models
tcplAddModel(pars = p1[[3]][[i]], modl = "hill")
}
#Add BMR
abline(h=bmr_thresh, lty = "dashed", col = "red")
#Add BMDU
{if(temp_bmds$ciu_POD != temp_bmds$highest_conc)abline(v=temp_bmds$ciu_POD, col = "yellow")}
#Add BMDL
{if(temp_bmds$cil_POD != temp_bmds$highest_conc)abline(v=temp_bmds$cil_POD, col = "purple")}
#Add BMD
{if(temp_bmds$median_POD != temp_bmds$highest_conc)abline(v=temp_bmds$median_POD, col = "blue")}
#Add Hit confidence
text(x = min(unique(p1_point$conc)+1), y = 0.8, paste0("Hit = ", temp_bmds$hit_confidence))

p <- recordPlot()

p_list[[chemicalNames[k]]] <- p

rm(p, p1, p1_point, temp_simulated, temp_bmds)

print(paste("Done", chemicalNames[k]))

}

rm(bmds_fitted, fitted_output_summary)
gc()
```

```{r alternative_save_the_resutls, include=FALSE}
save_png <- function(i) {
  png(
    paste0(
      "Output/Images/Affected_Rate/",
      chemicalNames[i],
      "_hill_fitted_AR_", ss, "_samples.png"
    ),
      units = "px",
      width = 325,
      height = 215,
      bg = "white"
  )
}

save_png(i=1)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=2)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=3)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=4)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=5)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=6)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=7)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=8)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=9)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=10)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=11)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=12)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=13)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=14)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=15)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=16)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=17)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=18)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=19)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=20)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=21)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=22)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=23)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=24)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=25)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=26)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=27)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=28)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

save_png(i=29)
p_list[[1]]
dev.off()
p_list[[1]] <- NULL

rm(p_list)
gc()

source("Functions/Save_fitted_model_images.R", local = TRUE)

rm(png_list, yleft, bottom)
gc()
```

```{r alternative_print_image, echo=FALSE}
knitr::include_graphics(
  paste0(
    "Output/Images/",
    "all_chems_hill_fitted_AR_",
    ss,
    "_samples.png"
  )
)
```
